python/voice-live-avatar/requirements.txt
azure-ai-voicelive[aiohttp]>=1.0.0
azure-identity
azure-core
azure-cosmos>=4.7.0
aiohttp
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
aiofiles
python-dotenv
